Add files via upload
This project was a banking application and it was developed using Python and the Pycharm IDE. This project gave an insight to how a customer can open an account, deposit/withdraw money and close an account. There was also a report of all the customers and their account balance as well as which customer held the most money in their account and also the total balance that was in the bank. There was an array of validation that was coded into each function so the program wouldn't crash as well as providing optimum efficiency for example if a customer is withdrawing an specific amount from their account and they didn't have enough money inside their accounts the program will let them know they have insufficient funds in their account.